Small improvements on the visualization code

+++ src/draw/draw.h
@@ -10,8 +10,8 @@
 
 struct draw_config {
     struct grid *grid_to_draw;
-    double max_v;
-    double min_v;
+    float max_v;
+    float min_v;
     bool grid_only;
     bool grid_lines;
     bool simulating;
@@ -20,7 +20,6 @@ struct draw_config {
     float time;
     float final_time;
     float dt;
-
 
     long solver_time;
     long ode_total_time;
